+++ CHANGELOG.md
@@ -1,7 +1,8 @@
-# v0.1.2
+# v0.2.0
 
 ## ✍️ Developer Notes
 - For help or other questions, concerns, etc. check out our Discord server: https://discord.gg/wp7mvmbkVb
 
 ## ✨ What's New
-- [Fabric] Fabric Loader 0.18.4 is now required.
+- Added ShapedRecipeBuilder#into(ItemStack).
+- Added ShapelessRecipeBuilder#into(ItemStack).